Extract dialog related code to dialog.go
Motivation: Have dialog component with minimal external dependencies.

Change: Define `Dialog` struct to explain what each fields are. To keep
all the logics untouched, it only references existing fields only.

Next plan: migrate remaining C code accessing those identified fields.

End goal: make sure all the memory access are migrated, then we can
safely move the fields definitions to dialog.go as well.
