+++ src/legacy/GAME2.c
@@ -327,17 +327,6 @@ int nox_xxx_playDialogFile_44D900(char* a1p, int a2) {
 	return 1;
 }
 
-//----- (0044D930) --------------------------------------------------------
-int sub_44D930() {
-	if (!dword_587000_122848) {
-		return 0;
-	}
-	if (dword_5d4594_830872 || dword_5d4594_831088) {
-		return 1;
-	}
-	return 0;
-}
-
 //----- (0044D960) --------------------------------------------------------
 void sub_44D960() { dword_587000_122848 = 0; }
